[LITTLE_FIX] : 🔧 데이터 연결 오류 수정
