+++ app/src/main/java/com/android/contactproject/contactlist/ContactListFragmentAdapter.kt
@@ -1,34 +1,31 @@
-package com.android.contactproject
+package com.android.contactproject.contactlist
 
-import android.content.Intent
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.android.contactproject.databinding.ContactListItemBinding
-import java.text.DecimalFormat
 
 class ContactListFragmentAdapter: RecyclerView.Adapter<ContactListFragmentAdapter.ViewHolder>() {
-    interface ItemLongClick {
-        fun onLongClick(view: View, position: Int )
+//    interface ItemLongClick {
+//        fun onLongClick(view: View, position: Int )
+//
+//    }
+//    interface ItemClick {
+//        fun onClick(view: View, position: Int )
+//    }
 
-    }
-    interface ItemClick {
-        fun onClick(view: View, position: Int )
-    }
 
+    private var list = mutableListOf<UserDataModel>()
 
-    private val list = ArrayList<UserDataModel>()
-
-    fun addItem(phoneNumberModel: UserDataModel) {
-        if (phoneNumberModel == null) {
-            return
-        }
-
-        list.add(phoneNumberModel)
-        notifyItemChanged(list.size - 1)
-    }
+//    fun addItem(phoneNumberModel: UserDataModel) {
+//        if (phoneNumberModel == null) {
+//            return
+//        }
+//
+//        list.add(phoneNumberModel)
+//        notifyItemChanged(list.size - 1)
+//    }
 
     override fun getItemCount(): Int {
         return list.size
@@ -40,11 +37,12 @@ class ContactListFragmentAdapter: RecyclerView.Adapter<ContactListFragmentAdapte
         )
     }
 
+
+
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = list[position]
-        val phNumber = list[position].ph
         holder.bind(item)
-        holder.phNumber.text = DecimalFormat("###-####-####").format(phNumber)
+
 
 
 
@@ -54,14 +52,22 @@ class ContactListFragmentAdapter: RecyclerView.Adapter<ContactListFragmentAdapte
     inner class ViewHolder(
         private val binding: ContactListItemBinding
     ) : RecyclerView.ViewHolder(binding.root) {
-        val phNumber = binding.phNumber
+
 
 
         fun bind(item: UserDataModel) = with(binding) {
            // 데이터
-            name.text = item.name
+            userName.text = item.name
+            phNumber.text = item.ph
+            profileImage.setImageResource(list[position].userImage)
+            mainLike.setImageResource(list[position].isLike)
+
         }
 
+    }
+    fun replace(newList: MutableList<UserDataModel>){
+        list = newList.toMutableList()
+        notifyDataSetChanged()
     }
 
 }
